autohighlight/requirements.txt
ffmpeg-python
tdvutil
tesserocr
opencv-contrib-python
pillow
